netextureforios: reconstruct +LoadTexture:Scale: at 0x32320
The header declared this class method but nothing defined it. It only
linked while -dead_strip elided gamesystem.o; now that a live call
reaches GameSystem, the missing symbol surfaces as a link failure.

The routine is the NSData sibling of C_TEXTURE::LoadFromUIImage and
follows the same shape. It builds a UIImage with initWithData: at
0x32374 and returns null when that is nil (the cbz at 0x3237c). It
takes .CGImage at 0x32398, reads the width and height at 0x323a4 and
0x323b0, then rounds each up to a power of two in the two shift loops
at 0x323b8 and 0x323cc. It draws the image into a zeroed RGBA8888
bitmap context, flipping the CTM first; the translate at 0x32454 uses
(double)imageHeight from the scvtf d9,w23 at 0x32444, not the
allocated height, which is what leaves the drawn region at the bottom
of the allocation.

The alpha test at 0x32494 is (alphaInfo - 1) unsigned-less-than 4, so
formats 1 through 4 keep the RGBA buffer as format 1 while 0, 5, and 6
fall into the tight 24-bit repack at 0x324b4 and upload as format 2.
The result goes to C_TEXTURE::CreateAndCache at 0x33d3c, which returns
the entry holding one reference; the callers in gamesystem.mm own that
reference and drop it through Release, as they already do.

Scale is a float, not the double the header carried. The method's type
encoding ends f24, and SetDataAndUpload stores the register it is
forwarded in with str s0 at 0x31f84, with no fcvt anywhere on the
path, so a double parameter would have narrowed the value the binary
passes through intact. The three call sites pass literals, so the
correction does not change them.

theos/Makefile needs no edit: it globs Project for sources with find,
so it picks the new translation unit up already.

+++ Project/GameSystem/src/OpenGL/neTextureForiOS.h
@@ -21,11 +21,16 @@ class C_TEXTURE;
 
 /**
  * @brief Decodes @p pData into a reference-counted engine texture at the given screen scale.
+ * The returned texture carries one reference, which the caller owns and drops through
+ * @c ne::C_TEXTURE::Release.
  * @param pData The encoded image data.
- * @param flScale The screen scale (1 or 2) the texture is decoded for.
+ * @param flScale The screen scale (1 or 2) the texture is decoded for. The method's type encoding
+ *        ends @c f24, and @c C_TEXTURE::SetDataAndUpload stores the forwarded register with
+ *        @c str @c s0, so this is a single-precision float rather than a double.
  * @return The loaded texture, or @c nullptr when the data could not be decoded.
+ * @ghidraAddress 0x32320
  */
-+ (nullable ne::C_TEXTURE *)LoadTexture:(nullable NSData *)pData Scale:(double)flScale;
++ (nullable ne::C_TEXTURE *)LoadTexture:(nullable NSData *)pData Scale:(float)flScale;
 
 @end
